code clean

+++ Render.cpp
@@ -82,7 +82,7 @@ void Render::DrawTexture(GPU_Image* texture, vec2f postion, vec2f scale, float a
 	GPU_BlitTransform(texture, NULL, _instance->_screenTexture_target, postion.x, postion.y, angle, scale.x, scale.y);
 }
 void Render::DrawSprite(Sprite* sprite, vec2f postion, int frame) {
-	DrawSprite_ex(sprite, postion.x, postion.y, frame, 1.0f, 1.0f, sprite->GetCenterX(), sprite->GetCenterY(), 0.0f);
+	DrawSprite_ex(sprite, postion.x, postion.y, frame, 1.0f, 1.0f, (float)sprite->GetCenterX(), (float)sprite->GetCenterY(), 0.0f);
 }
 
 void Render::DrawSprite_ex(Sprite* sprite, float posX, float posY, int frame, float scaleX, float scaleY, float centerX, float centerY, float angle) {
